New Main Mail View structures
MailListViewController internal structures (conversations by account and conversations by date) have been moved to external classes.  This is safer, as object types are better enforced, provides great debugging via description() methods, and makes the MailListViewController code more understandable.

+++ CocoaMail/Backend/DataBaseCore/Models/Accounts.m
@@ -575,9 +575,9 @@ typedef NSMutableArray<Conversation*> CCMMutableConversationArray;
                 DDLogInfo(@"Not All Account AND Is Not Connected.");
                 
                 [[ImapSync doLogin:acnt.user] subscribeError:^(NSError *error) {
-                    DDLogError(@"Account[%ld] connection failed, error = %@",acntIndex,error);
+                    DDLogError(@"Account[%ld] connection failed, error = %@",(long)acntIndex,error);
                 } completed:^{
-                    DDLogInfo(@"Account[%ld] connection success.",acntIndex);
+                    DDLogInfo(@"Account[%ld] connection success.",(long)acntIndex);
                 }];
                 
                 break;
